fix errors and warnings - further cleanup codebase
