+++ examples/src/data/images.js
@@ -1,42 +1,50 @@
 export default [
-    <img
-      src="https://images.ctfassets.net/m9id7be8gm8y/IU8tFfUW4ecCekM48Oa6y/30d4f82e4aeca32f26460641bea29bb1/Ilse.jpg?w=1240&h=1240&q=50&fm=webp"
-      alt="Album one"
-      data-action="https://github.com/easybird/react-cover-carousel"
-    />,
-    <img
-      src="https://images.ctfassets.net/m9id7be8gm8y/4jrg2hatXis8eOeIAkiuCs/b89146812b58b0b85842336dde4d5f17/Goldset.jpg?w=1240&h=1240&q=50&fm=webp"
-      alt="Album two"
-      data-action="https://github.com/easybird/react-cover-carousel"
-    />,
-    <img
-      src="https://images.ctfassets.net/m9id7be8gm8y/7zhDVJR79eaISS4ikQ8aIq/2fdd630d6077ac0ae2cc3c7ecb68a706/1-1-2.jpg?w=1240&h=1240&q=50&fm=webp"
-      alt="Album three"
-      data-action="https://github.com/easybird/react-cover-carousel"
-    />,
-    <img
-      src="https://images.ctfassets.net/m9id7be8gm8y/NrdbVLzRgiawmeOEwYEyO/d4e1bb3bed93759f845f3ea963a23f1a/Foto_Home_scherper-8.jpg?w=1240&h=1240&q=50&fm=webp"
-      alt="Album four"
-      data-action="https://github.com/easybird/react-cover-carousel"
-    />,
-    <img
-      src="https://images.ctfassets.net/m9id7be8gm8y/1XHV8OkABGuo0oEO0EAkSU/4dd104409ffdc4ebc17b7f1a7cb03788/Foto_Home_scherper-4.jpg?w=1240&h=1240&q=50&fm=webp"
-      alt="Album one"
-      data-action="https://github.com/easybird/react-cover-carousel"
-    />,
-    <img
-      src="https://images.ctfassets.net/m9id7be8gm8y/5zbdJ1gd56MwE6eQeY4mg4/603942e156d7b697f3d475f4ca578ca2/driehoeksring.jpg?w=1240&h=1240&q=50&fm=webp"
-      alt="Album two"
-      data-action="https://github.com/easybird/react-cover-carousel"
-    />,
-    <img
-      src="https://images.ctfassets.net/m9id7be8gm8y/4O69uPoL3iSIsk4SOEocCE/2f71cac33588cbcc37909e4973b4d955/25006526_1948828251798697_5933606261269463040_n.jpg?w=1080&h=1080&q=50&fm=webp"
-      alt="Album three"
-      data-action="https://github.com/easybird/react-cover-carousel"
-    />,
-    <img
-      src="https://images.ctfassets.net/m9id7be8gm8y/5KY6HcXfTqGUm2qGykAwII/1b604ae281c40af25ed488c4424ac143/5-10-11-2.jpg?w=1240&h=1240&q=50&fm=webp"
-      alt="Album four"
-      data-action="https://github.com/easybird/react-cover-carousel"
-    />,
-  ];
+  <img
+    key="https://images.ctfassets.net/m9id7be8gm8y/IU8tFfUW4ecCekM48Oa6y/30d4f82e4aeca32f26460641bea29bb1/Ilse.jpg?w=1240&h=1240&q=50&fm=webp"
+    src="https://images.ctfassets.net/m9id7be8gm8y/IU8tFfUW4ecCekM48Oa6y/30d4f82e4aeca32f26460641bea29bb1/Ilse.jpg?w=1240&h=1240&q=50&fm=webp"
+    alt="Album one"
+    data-action="https://github.com/easybird/react-cover-carousel"
+  />,
+  <img
+    key="https://images.ctfassets.net/m9id7be8gm8y/4jrg2hatXis8eOeIAkiuCs/b89146812b58b0b85842336dde4d5f17/Goldset.jpg?w=1240&h=1240&q=50&fm=webp"
+    src="https://images.ctfassets.net/m9id7be8gm8y/4jrg2hatXis8eOeIAkiuCs/b89146812b58b0b85842336dde4d5f17/Goldset.jpg?w=1240&h=1240&q=50&fm=webp"
+    alt="Album two"
+    data-action="https://github.com/easybird/react-cover-carousel"
+  />,
+  <img
+    key="https://images.ctfassets.net/m9id7be8gm8y/7zhDVJR79eaISS4ikQ8aIq/2fdd630d6077ac0ae2cc3c7ecb68a706/1-1-2.jpg?w=1240&h=1240&q=50&fm=webp"
+    src="https://images.ctfassets.net/m9id7be8gm8y/7zhDVJR79eaISS4ikQ8aIq/2fdd630d6077ac0ae2cc3c7ecb68a706/1-1-2.jpg?w=1240&h=1240&q=50&fm=webp"
+    alt="Album three"
+    data-action="https://github.com/easybird/react-cover-carousel"
+  />,
+  <img
+    key="https://images.ctfassets.net/m9id7be8gm8y/NrdbVLzRgiawmeOEwYEyO/d4e1bb3bed93759f845f3ea963a23f1a/Foto_Home_scherper-8.jpg?w=1240&h=1240&q=50&fm=webp"
+    src="https://images.ctfassets.net/m9id7be8gm8y/NrdbVLzRgiawmeOEwYEyO/d4e1bb3bed93759f845f3ea963a23f1a/Foto_Home_scherper-8.jpg?w=1240&h=1240&q=50&fm=webp"
+    alt="Album four"
+    data-action="https://github.com/easybird/react-cover-carousel"
+  />,
+  <img
+    key="https://images.ctfassets.net/m9id7be8gm8y/1XHV8OkABGuo0oEO0EAkSU/4dd104409ffdc4ebc17b7f1a7cb03788/Foto_Home_scherper-4.jpg?w=1240&h=1240&q=50&fm=webp"
+    src="https://images.ctfassets.net/m9id7be8gm8y/1XHV8OkABGuo0oEO0EAkSU/4dd104409ffdc4ebc17b7f1a7cb03788/Foto_Home_scherper-4.jpg?w=1240&h=1240&q=50&fm=webp"
+    alt="Album one"
+    data-action="https://github.com/easybird/react-cover-carousel"
+  />,
+  <img
+    key="https://images.ctfassets.net/m9id7be8gm8y/5zbdJ1gd56MwE6eQeY4mg4/603942e156d7b697f3d475f4ca578ca2/driehoeksring.jpg?w=1240&h=1240&q=50&fm=webp"
+    src="https://images.ctfassets.net/m9id7be8gm8y/5zbdJ1gd56MwE6eQeY4mg4/603942e156d7b697f3d475f4ca578ca2/driehoeksring.jpg?w=1240&h=1240&q=50&fm=webp"
+    alt="Album two"
+    data-action="https://github.com/easybird/react-cover-carousel"
+  />,
+  <img
+    key="https://images.ctfassets.net/m9id7be8gm8y/4O69uPoL3iSIsk4SOEocCE/2f71cac33588cbcc37909e4973b4d955/25006526_1948828251798697_5933606261269463040_n.jpg?w=1080&h=1080&q=50&fm=webp"
+    src="https://images.ctfassets.net/m9id7be8gm8y/4O69uPoL3iSIsk4SOEocCE/2f71cac33588cbcc37909e4973b4d955/25006526_1948828251798697_5933606261269463040_n.jpg?w=1080&h=1080&q=50&fm=webp"
+    alt="Album three"
+    data-action="https://github.com/easybird/react-cover-carousel"
+  />,
+  <img
+    key="https://images.ctfassets.net/m9id7be8gm8y/5KY6HcXfTqGUm2qGykAwII/1b604ae281c40af25ed488c4424ac143/5-10-11-2.jpg?w=1240&h=1240&q=50&fm=webp"
+    src="https://images.ctfassets.net/m9id7be8gm8y/5KY6HcXfTqGUm2qGykAwII/1b604ae281c40af25ed488c4424ac143/5-10-11-2.jpg?w=1240&h=1240&q=50&fm=webp"
+    alt="Album four"
+    data-action="https://github.com/easybird/react-cover-carousel"
+  />,
+];
