+++ src/styles/coverCarousel.css
@@ -1,0 +1,20 @@
+.ReactCoverCarousel {
+    position: relative;
+    width: 100%;
+    height: 100%;
+    margin: 0;
+  }
+  
+  .stage {
+    height: 100%;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    margin: 0;
+    transform-style: preserve-3d;
+    perspective: 500px;
+  }
+  
+  .preloader {
+    display: hidden;
+  }
